% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{cohort_boxplot}
\alias{cohort_boxplot}
\title{Generate boxplot.}
\usage{
cohort_boxplot(blood.module, orderByTissue, orderByModule,
  cohort.name = "all", patient.ids = NULL)
}
\description{
Generate boxplot.
}
