% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{commonEnrichmentScoreGenes}
\alias{commonEnrichmentScoreGenes}
\title{Get common genes from er analysis with genelist and module}
\usage{
commonEnrichmentScoreGenes(tissue, module, genelist)
}
\description{
Get common genes from er analysis with genelist and module
}

