% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{comparisonAnalyses}
\alias{comparisonAnalyses}
\title{Compute all 4 different analyses for modules from two tissues.}
\usage{
comparisonAnalyses(tissueA, tissueB, moduleA, moduleB, cohort = "all")
}
\arguments{
\item{tissueA}{is the first tissue}

\item{tissueB}{is the second tissue}

\item{moduleA}{is a module from the first tissue}

\item{moduleB}{is a module from the second tissue}
}
\description{
Compute all 4 different analyses for modules from two tissues.
}
