% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{computeEigengenes}
\alias{computeEigengenes}
\title{Computes module eigengenes for modules in the given tissues}
\usage{
computeEigengenes(tissues)
}
\arguments{
\item{tissues}{is a vector of tissues}
}
\value{
MEs object with module eigengenes for each tissue
}
\description{
Computes module eigengenes for modules in the given tissues
}

