% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/app.R
\name{eigengeneClinicalRelation}
\alias{eigengeneClinicalRelation}
\title{Computes correlation between eigengenes and quantitative variables
and anova between eigengenes and qualitative variables.}
\usage{
eigengeneClinicalRelation(tissue)
}
\arguments{
\item{tissue}{is the tissue from which to get the eigengenes}
}
\description{
Computes correlation between eigengenes and quantitative variables
and anova between eigengenes and qualitative variables.
}

