% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{eigengeneCorrelation}
\alias{eigengeneCorrelation}
\title{Calculates eigengene correlations between tissue A and tissue B. Returns
the p-values.}
\usage{
eigengeneCorrelation(tissueA, tissueB)
}
\description{
Calculates eigengene correlations between tissue A and tissue B. Returns
the p-values.
}

