% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/app.R
\name{eigengeneHypergeometricTest}
\alias{eigengeneHypergeometricTest}
\title{Run gene overlap test between eigengenes in tissueA and tissueB. Returns
p values}
\usage{
eigengeneHypergeometricTest(tissueA, tissueB)
}
\arguments{
\item{tissueA}{is the first tissue we are interested in, e.g. "blood"}

\item{tissueB}{is the other, e.g. "biopsy"}
}
\description{
Run gene overlap test between eigengenes in tissueA and tissueB. Returns
p values
}

