% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{geneOverlapTest}
\alias{geneOverlapTest}
\title{Compute gene overlap between genes from two tissues}
\usage{
geneOverlapTest(tissueA = "blood", tissueB = "biopsy")
}
\description{
Compute gene overlap between genes from two tissues
}

