% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{getAllGenes}
\alias{getAllGenes}
\title{Returns a list of all genes found in  all modules across all tissues.}
\usage{
getAllGenes()
}
\value{
vector of gene names
}
\description{
Returns a list of all genes found in  all modules across all tissues.
}
