% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/app.R
\name{getAllGenesAndModules}
\alias{getAllGenesAndModules}
\title{Retrieves all genes and the modules they participate in.}
\usage{
getAllGenesAndModules()
}
\description{
Retrieves all genes and the modules they participate in.
}

