% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/app.R
\name{getAllModules}
\alias{getAllModules}
\title{Get all modules a specific gene is found in.}
\usage{
getAllModules(gene)
}
\arguments{
\item{gene}{the interesting gene}
}
\description{
Get all modules a specific gene is found in.
}

