% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{getCommonGenes}
\alias{getCommonGenes}
\title{Get common genes for given tissue, module, geneset and optionally status}
\usage{
getCommonGenes(tissue, module, geneset, status = "updn.common")
}
\arguments{
\item{geneset}{is the geneset, from any of the msgidb sets}

\item{status}{is the status, up.common, dn.common or default updn.common}
}
\description{
Get common genes for given tissue, module, geneset and optionally status
}
