% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{getCommonGOTermGenes}
\alias{getCommonGOTermGenes}
\title{Get the common genes from the go term anlysis for a specific
tissue, module and GO term id.}
\usage{
getCommonGOTermGenes(tissue, module, gotermID)
}
\arguments{
\item{gotermID}{is the GO term id, e.g. "GO:0070848"}
}
\description{
Get the common genes from the go term anlysis for a specific
tissue, module and GO term id.
}
