% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{getEnrichmentForTissue}
\alias{getEnrichmentForTissue}
\title{Get enrichment scores for all modules}
\usage{
getEnrichmentForTissue(tissue, genesets = c(1))
}
\arguments{
\item{tissue}{is the tissue we're interested in, e.g. blood}

\item{genesets}{is a vector of the genesets we want to look at, default is the first...}
}
\description{
Get enrichment scores for all modules
}

