% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/app.R
\name{getEnrichmentScores}
\alias{getEnrichmentScores}
\title{Get enrichment scores for specific module and tissue}
\usage{
getEnrichmentScores(tissue, module)
}
\description{
Get enrichment scores for specific module and tissue
}

