% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{getGeneList}
\alias{getGeneList}
\title{Get a list of genes for a specific module and tissue.}
\usage{
getGeneList(tissue, module, cohort = "all")
}
\arguments{
\item{tissue}{is the tissue we're interested in}

\item{module}{is the module we want to get the genes from}
}
\value{
matrix with columns: gene name, up.dn, cor
}
\description{
Get a list of genes for a specific module and tissue.
}

