% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{getGOScoresForTissue}
\alias{getGOScoresForTissue}
\title{Get all scores for a specific term in a tissue}
\usage{
getGOScoresForTissue(tissue, term)
}
\description{
Get all scores for a specific term in a tissue
}

