% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{getGOTerms}
\alias{getGOTerms}
\title{Get go terms for the given module and tissue. Possible to specify
which specific terms you're interested in.}
\usage{
getGOTerms(tissue, module, terms = c())
}
\arguments{
\item{tissue}{is the tissue, e.g. blood}

\item{module}{is the module, e.g. blue, pink etc.}

\item{terms}{are a vector GO terms we're interested in, default is all, given as a vector.}
}
\description{
Get go terms for the given module and tissue. Possible to specify
which specific terms you're interested in.
}

