% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{getModules}
\alias{getModules}
\title{Returns a list of modules found for the given tissue}
\usage{
getModules(tissue)
}
\arguments{
\item{tissue}{tissue we want to retrieve modules for}
}
\value{
vector of module names
}
\description{
Returns a list of modules found for the given tissue
}
