% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/app.R
\name{heatmap}
\alias{heatmap}
\title{Generate heatmap plot for the given tissue and module. If the heatmap
already exists, it finds the appropriate png file where it is supposed
to store a new one, it returns this file. This heat map function
generates both a png and a pdf. All plots are stored in the path
given by 'imgpath'}
\usage{
heatmap(tissue, module)
}
\arguments{
\item{tissue}{we want to look at}

\item{module}{the module we want to generate a heetmap for}
}
\description{
Generate heatmap plot for the given tissue and module. If the heatmap
already exists, it finds the appropriate png file where it is supposed
to store a new one, it returns this file. This heat map function
generates both a png and a pdf. All plots are stored in the path
given by 'imgpath'
}
\examples{
heatmap("blood", "green")
heatmap("biopsy", "blue")
}

