\name{mixt-package}
\alias{mixt-package}
\alias{mixt}
\docType{package}
\title{
\packageTitle{mixt}
}
\description{
\packageDescription{mixt}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{mixt}
\packageIndices{mixt}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{mixt}

Maintainer: \packageMaintainer{mixt}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
