% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{moduleHypergeometricTest}
\alias{moduleHypergeometricTest}
\title{Run gene overlap test between eigengenes in tissueA and tissueB. Returns 
p values}
\usage{
moduleHypergeometricTest(tissueA, tissueB)
}
\arguments{
\item{tissueA}{is the first tissue we are interested in, e.g. "blood"}

\item{tissueB}{is the other, e.g. "biopsy"}
}
\description{
Run gene overlap test between eigengenes in tissueA and tissueB. Returns 
p values
}

