% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{patientRankCorrelation}
\alias{patientRankCorrelation}
\title{Compute correlation between patient rank in two tissues}
\usage{
patientRankCorrelation(tissueA = "blood", tissueB = "biopsy")
}
\arguments{
\item{subtypes}{is a vector of subtypes}
}
\description{
Compute correlation between patient rank in two tissues
}

