% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{userEnrichmentScores}
\alias{userEnrichmentScores}
\title{Get p-values from the hypergeometric test between an arbirtrary gene list
and the modules. User must specify tissue and genelist}
\usage{
userEnrichmentScores(tissue, genelist, cohort = "all")
}
\arguments{
\item{tissue}{is the tissue we're interested in, e.g. "blood"}

\item{genelist}{is the genelist as a vector.}
}
\description{
Get p-values from the hypergeometric test between an arbirtrary gene list
and the modules. User must specify tissue and genelist
}

