% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{getAllGenesAndModules}
\alias{getAllGenesAndModules}
\title{Retrieves an overview of all genes and the modules they participate in.}
\usage{
getAllGenesAndModules(cohort = "all")
}
\description{
Retrieves an overview of all genes and the modules they participate in.
}
