% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{getAllModules}
\alias{getAllModules}
\title{Get all modules a specific gene is found in.}
\usage{
getAllModules(gene)
}
\arguments{
\item{gene}{the interesting gene}
}
\value{
vector with module name from blood and module name from biopsy
}
\description{
Get all modules a specific gene is found in.
}
