% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/app.R
\name{getGeneList}
\alias{getGeneList}
\title{Get a list of genes for a specific module and tissue.}
\usage{
getGeneList(tissue, module)
}
\arguments{
\item{tissue}{is the tissue we're interested in}

\item{module}{is the module we want to get the genes from}
}
\description{
Get a list of genes for a specific module and tissue.
}

