% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{patientRankSum}
\alias{patientRankSum}
\title{Calculate patient rank sum scores for given tissues.}
\usage{
patientRankSum(tissueA = "blood", tissueB = "biopsy", cohort = "all")
}
\description{
Calculate patient rank sum scores for given tissues.
}

